--- .idea/libraries/Maven__log4j_log4j_1_2_17.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: log4j:log4j:1.2.17">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/log4j/log4j/1.2.17/log4j-1.2.17.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/log4j/log4j/1.2.17/log4j-1.2.17-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/log4j/log4j/1.2.17/log4j-1.2.17-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
