--- src/test/java/com/blackNecoProject/testng/commons/testHelper/FailureRetryListener.java
@@ -1,19 +1,0 @@
-package com.blackNecoProject.testng.commons.testHelper;
-
-import org.testng.IAnnotationTransformer;
-import org.testng.IRetryAnalyzer;
-import org.testng.annotations.ITestAnnotation;
-
-import java.lang.reflect.Constructor;
-import java.lang.reflect.Method;
-
-public class FailureRetryListener implements IAnnotationTransformer {
-
-    public void transform(ITestAnnotation testAnnotation, Class testClass,
-                          Constructor testConstructor, Method testMethod){
-        IRetryAnalyzer retry = testAnnotation.getRetryAnalyzer();
-        if (retry == null)
-            testAnnotation.setRetryAnalyzer(RetryAnalyzer.class);
-    }
-
-}
