--- src/test/java/com/blackNecoProject/testng/commons/testHelper/RetryAnalyzer.java
@@ -1,19 +1,0 @@
-package com.blackNecoProject.testng.commons.testHelper;
-
-import org.testng.IRetryAnalyzer;
-import org.testng.ITestResult;
-
-public class RetryAnalyzer implements IRetryAnalyzer {
-    int retryAttemptsCounter = 0;
-    int maxRetryLimit = 1;
-
-    public boolean retry(ITestResult result){
-        if (!result.isSuccess()) {
-            if (retryAttemptsCounter < maxRetryLimit) {
-                retryAttemptsCounter++;
-                return true;
-            }
-        }
-        return false;
-    }
-}
