+++ src/test/java/commons/Config.java
@@ -1,8 +1,8 @@
-package com.blackNecoProject.testng.commons;
+package commons;
 
 public class Config {
     //Valid credential
-    public static String EMAIL ="bedove69@meta.ua";
+    public static String Email ="bedove69@meta.ua";
     public static String USERNAME = "mgr123";
     public static String PASSWORD = "mgr!23";
