--- src/test/java/testNG/ExpediteTests.xml
@@ -1,29 +1,0 @@
-<!--<?xml version="1.0" encoding="UTF-8"?>-->
-<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd">
-<suite name="Expedite Tests" verbose="1" preserve-order="true">
-    <listeners>
-        <listener class-name="com.blackNecoProject.testng.commons.testHelper.FailureRetryListener"/>
-        <listener class-name="com.blackNecoProject.testng.commons.testHelper.TestListener"/>
-    </listeners>
-
-    <test name="Mail Test">
-        <classes>
-            <class name="com.blackNecoProject.testng.tests.MailPageTest">
-                <methods>
-                    <include name="getAccessToDemoSite"/>
-                </methods>
-            </class>
-        </classes>
-    </test>
-
-    <test name="Login test">
-        <classes>
-            <class name="com.blackNecoProject.testng.tests.LoginTests">
-                <methods>
-                    <include name="userCanLoginSuccessfully"/>
-                </methods>
-            </class>
-        </classes>
-    </test>
-
-</suite>
