+++ src/test/java/tests/BaseTest.java
@@ -1,0 +1,25 @@
+package tests;
+
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.chrome.ChromeDriver;
+import org.testng.annotations.AfterTest;
+import org.testng.annotations.BeforeTest;
+
+import java.util.concurrent.TimeUnit;
+
+public class BaseTest {
+    protected WebDriver driver;
+
+    @BeforeTest
+    public void setup(){
+        driver = new ChromeDriver();
+        driver.manage().timeouts().implicitlyWait(10, TimeUnit.SECONDS);
+        driver.manage().window().maximize();
+        driver.get("http://demo.guru99.com/V4/");
+    }
+
+    @AfterTest
+    public void turnDown(){
+        driver.quit();
+    }
+}
