+++ src/test/java/tests/LoginTests.java
@@ -1,0 +1,33 @@
+package tests;
+
+import org.testng.Assert;
+import org.testng.annotations.Test;
+import pages.HomePage;
+import pages.LoginPage;
+
+public class LoginTests extends BaseTest {
+
+    private LoginPage objLogin;
+    private HomePage objHomePage;
+
+    /**
+     * This test case will login in http://demo.guru99.com/V4/
+     * Verify login page title as guru99 bank
+     * Login to application
+     * Verify the home page using Dashboard message
+     */
+    @Test(description = "Test Case 1 - User can login successfully")
+    public void userCanLoginSuccessfully() {
+        //Create Login Page object
+        objLogin = new LoginPage(driver);
+        //Verify login page title
+        String loginPageTitle = objLogin.getLoginTitle();
+        Assert.assertTrue(loginPageTitle.toLowerCase().contains("guru99 bank"));
+        //login to application
+        objLogin.loginToGuru99("mgr123", "mgr!23");
+        // go the next page
+        objHomePage = new HomePage(driver);
+        //Verify home page
+        Assert.assertTrue(objHomePage.getHomePageDashboardUserName().toLowerCase().contains("manger id : mgr123"));
+    }
+}
